Change names for autogenerated layouts
