+++ surface_sim/layouts/__init__.py
@@ -1,9 +1,9 @@
 from .layout import Layout
 from .library import (
-    rot_surf_code,
-    rot_surf_code_rectangle,
-    unrot_surf_code,
-    unrot_surf_code_rectangle,
+    rot_surface_code,
+    rot_surface_code_rectangle,
+    unrot_surface_code,
+    unrot_surface_code_rectangle,
 )
 from .plotter import plot
 from .util import set_coords
@@ -11,10 +11,10 @@ from .operations import check_overlap_layouts
 
 __all__ = [
     "Layout",
-    "rot_surf_code",
-    "rot_surf_code_rectangle",
-    "unrot_surf_code",
-    "unrot_surf_code_rectangle",
+    "rot_surface_code",
+    "rot_surface_code_rectangle",
+    "unrot_surface_code",
+    "unrot_surface_code_rectangle",
     "plot",
     "set_coords",
     "check_overlap_layouts",
