+++ surface_sim/layouts/library/__init__.py
@@ -1,9 +1,9 @@
-from .rot_surface_code import rot_surf_code, rot_surf_code_rectangle
-from .unrot_surface_code import unrot_surf_code, unrot_surf_code_rectangle
+from .rot_surface_codes import rot_surface_code, rot_surface_code_rectangle
+from .unrot_surface_codes import unrot_surface_code, unrot_surface_code_rectangle
 
 __all__ = [
-    "rot_surf_code",
-    "rot_surf_code_rectangle",
-    "unrot_surf_code",
-    "unrot_surf_code_rectangle",
+    "rot_surface_code",
+    "rot_surface_code_rectangle",
+    "unrot_surface_code",
+    "unrot_surface_code_rectangle",
 ]
