+++ surface_sim/layouts/library/rot_surface_codes.py
@@ -1,6 +1,6 @@
 from collections import defaultdict
 from functools import partial
-from itertools import count
+from itertools import count, cycle, product
 
 from ..layout import Layout
 from .util import is_valid, invert_shift, _check_distance
@@ -29,17 +29,13 @@ def get_data_index(row: int, col: int, col_size: int, start_ind: int = 1) -> int
     int
         The index of the data qubit.
     """
-    if row % 2 == 1:
-        row -= 1
-        col += col_size
-
     row_ind = row // 2
     col_ind = col // 2
-    index = start_ind + row_ind * col_size + col_ind
+    index = start_ind + (row_ind * col_size) + col_ind
     return index
 
 
-def shift_direction(shift: tuple[int, int]) -> str:
+def shift_direction(row_shift: int, col_shift: int) -> str:
     """Translates a row and column shift to a direction.
 
     Parameters
@@ -54,23 +50,36 @@ def shift_direction(shift: tuple[int, int]) -> str:
     str
         The direction.
     """
-    if shift == (0, 1):
-        return "north"
-    elif shift == (0, -1):
-        return "south"
-    elif shift == (1, 0):
-        return "east"
-    elif shift == (-1, 0):
-        return "west"
-    else:
-        raise ValueError("The shift does not correspond to a known direction.")
+    ver_direction = "north" if row_shift > 0 else "south"
+    hor_direction = "east" if col_shift > 0 else "west"
+    direction = f"{ver_direction}_{hor_direction}"
+    return direction
 
 
-def unrot_surf_code_rectangle(
+def add_missing_neighbours(neighbor_data: dict) -> None:
+    """Adds None for missing neighbours in the neighbor data.
+    Note that this modifies the dictionary in place.
+
+    Parameters
+    ----------
+    neighbor_data
+        The neighbor data - a dictionary that contains information
+        about the layout connectivity (defined on a square grid,
+        connection run diagonally).
+    """
+    directions = ["north_east", "north_west", "south_east", "south_west"]
+    for neighbors in neighbor_data.values():
+        for direction in directions:
+            if direction not in neighbors:
+                neighbors[direction] = None
+    return None
+
+
+def rot_surface_code_rectangle(
     distance_x: int,
     distance_z: int,
     logical_qubit_label: str = "L0",
-    init_point: tuple[int | float, int | float] = (0, 0),
+    init_point: tuple[int | float, int | float] = (1, 1),
     init_data_qubit_id: int = 1,
     init_zanc_qubit_id: int = 1,
     init_xanc_qubit_id: int = 1,
@@ -137,19 +146,19 @@ def unrot_surf_code_rectangle(
             f"but {type(init_xanc_qubit_id)} was given."
         )
 
-    name = f"Unrotated dx-{distance_x} dz-{distance_z} surface code layout."
-    code = "unrotated_surface_code"
+    name = f"Rotated dx-{distance_x} dz-{distance_z} surface code layout."
+    code = "rotated_surface_code"
     description = None
 
+    freq_order = ["low", "mid", "high"]
+
     int_order = dict(
-        x_type=["north", "west", "east", "south"],
-        z_type=["north", "east", "west", "south"],
+        x_type=["north_east", "north_west", "south_east", "south_west"],
+        z_type=["north_east", "south_east", "north_west", "south_west"],
     )
 
     log_z = [f"D{i+1 + init_data_qubit_id}" for i in range(distance_z)]
-    log_x = [
-        f"D{i*(2*distance_z - 1)+1 + init_data_qubit_id}" for i in range(distance_x)
-    ]
+    log_x = [f"D{i*distance_z+1 + init_data_qubit_id}" for i in range(distance_x)]
 
     layout_setup = dict(
         name=name,
@@ -158,6 +167,7 @@ def unrot_surf_code_rectangle(
         description=description,
         distance_x=distance_x,
         distance_z=distance_z,
+        freq_order=freq_order,
         interaction_order=int_order,
         log_z={logical_qubit_label: log_z},
         log_x={logical_qubit_label: log_x},
@@ -165,70 +175,103 @@ def unrot_surf_code_rectangle(
     if distance_x == distance_z:
         layout_setup["distance"] = distance_z
 
-    col_size = 2 * distance_z - 1
-    row_size = 2 * distance_x - 1
+    col_size = 2 * distance_z + 1
+    row_size = 2 * distance_x + 1
     data_indexer = partial(
-        get_data_index, col_size=col_size, start_ind=init_data_qubit_id
+        get_data_index, col_size=distance_z, start_ind=init_data_qubit_id
     )
     valid_coord = partial(is_valid, max_size_col=col_size, max_size_row=row_size)
 
-    nbr_shifts = [(0, 1), (0, -1), (1, 0), (-1, 0)]
+    pos_shifts = (1, -1)
+    nbr_shifts = tuple(product(pos_shifts, repeat=2))
 
     layout_data = []
     neighbor_data = defaultdict(dict)
+    freq_seq = cycle(("low", "high"))
     ind = init_ind
 
+    # change initial point because by default the code places the "D1" qubit
+    # in the (1,1) point.
+    init_point = (init_point[0] - 1, init_point[1] - 1)
+
+    for row in range(1, row_size, 2):
+        freq_group = next(freq_seq)
+        for col in range(1, col_size, 2):
+            index = data_indexer(row, col)
+
+            qubit_info = dict(
+                qubit=f"D{index}",
+                role="data",
+                coords=(row + init_point[0], col + init_point[1]),
+                freq_group=freq_group,
+                stab_type=None,
+                ind=ind,
+            )
+            layout_data.append(qubit_info)
+
+            ind += 1
+
     x_index = count(start=init_xanc_qubit_id)
+    for row in range(0, row_size, 2):
+        for col in range(2 + row % 4, col_size - 1, 4):
+            anc_qubit = f"X{next(x_index)}"
+            qubit_info = dict(
+                qubit=anc_qubit,
+                role="anc",
+                coords=(row + init_point[0], col + init_point[1]),
+                freq_group="mid",
+                stab_type="x_type",
+                ind=ind,
+            )
+            layout_data.append(qubit_info)
+
+            ind += 1
+
+            for row_shift, col_shift in nbr_shifts:
+                data_row, data_col = row + row_shift, col + col_shift
+                if not valid_coord(data_row, data_col):
+                    continue
+                data_index = data_indexer(data_row, data_col)
+                data_qubit = f"D{data_index}"
+
+                direction = shift_direction(row_shift, col_shift)
+                neighbor_data[anc_qubit][direction] = data_qubit
+
+                inv_shifts = invert_shift(row_shift, col_shift)
+                inv_direction = shift_direction(*inv_shifts)
+                neighbor_data[data_qubit][inv_direction] = anc_qubit
+
     z_index = count(start=init_zanc_qubit_id)
-    for row in range(row_size):
-        for col in range(col_size):
-            role = "data" if (row + col) % 2 == 0 else "anc"
+    for row in range(2, row_size - 1, 2):
+        for col in range(row % 4, col_size, 4):
+            anc_qubit = f"Z{next(z_index)}"
+            qubit_info = dict(
+                qubit=anc_qubit,
+                role="anc",
+                coords=(row + init_point[0], col + init_point[1]),
+                freq_group="mid",
+                stab_type="z_type",
+                ind=ind,
+            )
+            layout_data.append(qubit_info)
 
-            if role == "data":
-                index = data_indexer(row, col)
+            ind += 1
 
-                qubit_info = dict(
-                    qubit=f"D{index}",
-                    role="data",
-                    coords=(row + init_point[0], col + init_point[1]),
-                    stab_type=None,
-                    ind=ind,
-                )
-                layout_data.append(qubit_info)
+            for row_shift, col_shift in nbr_shifts:
+                data_row, data_col = row + row_shift, col + col_shift
+                if not valid_coord(data_row, data_col):
+                    continue
+                data_index = data_indexer(data_row, data_col)
+                data_qubit = f"D{data_index}"
 
-                ind += 1
+                direction = shift_direction(row_shift, col_shift)
+                neighbor_data[anc_qubit][direction] = data_qubit
 
-            else:
-                stab_type = "x_type" if row % 2 == 0 else "z_type"
-                if stab_type == "x_type":
-                    anc_qubit = f"X{next(x_index)}"
-                else:
-                    anc_qubit = f"Z{next(z_index)}"
+                inv_shifts = invert_shift(row_shift, col_shift)
+                inv_direction = shift_direction(*inv_shifts)
+                neighbor_data[data_qubit][inv_direction] = anc_qubit
 
-                qubit_info = dict(
-                    qubit=anc_qubit,
-                    role="anc",
-                    coords=(row + init_point[0], col + init_point[1]),
-                    stab_type=stab_type,
-                    ind=ind,
-                )
-                layout_data.append(qubit_info)
-
-                ind += 1
-
-                for row_shift, col_shift in nbr_shifts:
-                    data_row, data_col = row + row_shift, col + col_shift
-                    if not valid_coord(data_row, data_col):
-                        continue
-                    data_index = data_indexer(data_row, data_col)
-                    data_qubit = f"D{data_index}"
-
-                    direction = shift_direction((row_shift, col_shift))
-                    neighbor_data[anc_qubit][direction] = data_qubit
-
-                    inv_shifts = invert_shift(row_shift, col_shift)
-                    inv_direction = shift_direction(inv_shifts)
-                    neighbor_data[data_qubit][inv_direction] = anc_qubit
+    add_missing_neighbours(neighbor_data)
 
     for qubit_info in layout_data:
         qubit = qubit_info["qubit"]
@@ -239,16 +282,16 @@ def unrot_surf_code_rectangle(
     return layout
 
 
-def unrot_surf_code(
+def rot_surface_code(
     distance: int,
     logical_qubit_label: str = "L0",
-    init_point: tuple[int | float, int | float] = (0, 0),
+    init_point: tuple[int | float, int | float] = (1, 1),
     init_data_qubit_id: int = 1,
     init_zanc_qubit_id: int = 1,
     init_xanc_qubit_id: int = 1,
     init_ind: int = 0,
 ) -> Layout:
-    """Generates an unrotated surface code layout.
+    """Generates a rotated surface code layout.
 
     Parameters
     ----------
@@ -258,7 +301,7 @@ def unrot_surf_code(
         Label for the logical qubit, by default ``"L0"``.
     init_point
         Coordinates for the bottom left (i.e. southest west) data qubit.
-        By default ``(1, 1)``.
+        By default ``1``, so the label is ``"D1"``.
     init_data_qubit_id
         Index for the bottom left (i.e. southest west) data qubit.
         By default ``1``, so the label is ``"D1"``.
@@ -276,7 +319,7 @@ def unrot_surf_code(
     Layout
         The layout of the code.
     """
-    return unrot_surf_code_rectangle(
+    return rot_surface_code_rectangle(
         distance_x=distance,
         distance_z=distance,
         logical_qubit_label=logical_qubit_label,
