+++ tests/experiments/test_rot_surface_code_css_pipelined.py
@@ -1,6 +1,6 @@
 import stim
 
-from surface_sim.layouts import rot_surf_code, rot_surf_code_rectangle
+from surface_sim.layouts import rot_surface_code, rot_surface_code_rectangle
 
 from surface_sim.experiments.rot_surface_code_css_pipelined import (
     memory_experiment,
@@ -12,7 +12,7 @@ from surface_sim.log_gates.rot_surface_code_css import set_trans_s
 
 
 def test_memory_experiment():
-    layout = rot_surf_code(distance=3)
+    layout = rot_surface_code(distance=3)
     qubit_ids = {q: i for i, q in enumerate(layout.get_qubits())}
     anc_coords = {q: layout.get_coords([q])[0] for q in layout.get_qubits(role="anc")}
     model = NoiselessModel(qubit_ids)
@@ -48,7 +48,7 @@ def test_memory_experiment():
 
 
 def test_repeated_s_experiment():
-    layout = rot_surf_code_rectangle(distance_z=4, distance_x=3)
+    layout = rot_surface_code_rectangle(distance_z=4, distance_x=3)
     set_trans_s(layout, "D1")
     qubit_ids = {q: i for i, q in enumerate(layout.get_qubits())}
     anc_coords = {q: layout.get_coords([q])[0] for q in layout.get_qubits(role="anc")}
@@ -86,7 +86,7 @@ def test_repeated_s_experiment():
 
 
 def test_memory_experiment_anc_detectors():
-    layout = rot_surf_code(distance=3)
+    layout = rot_surface_code(distance=3)
     qubit_ids = {q: i for i, q in enumerate(layout.get_qubits())}
     model = NoiselessModel(qubit_ids)
     detectors = Detectors(layout.get_qubits(role="anc"), frame="1")
@@ -116,7 +116,7 @@ def test_memory_experiment_anc_detectors():
 
 
 def test_repeated_s_experiment_anc_detectors():
-    layout = rot_surf_code_rectangle(distance_z=4, distance_x=3)
+    layout = rot_surface_code_rectangle(distance_z=4, distance_x=3)
     set_trans_s(layout, "D1")
     qubit_ids = {q: i for i, q in enumerate(layout.get_qubits())}
     model = NoiselessModel(qubit_ids)
@@ -148,7 +148,7 @@ def test_repeated_s_experiment_anc_detectors():
 
 
 def test_memory_experiment_gauge_detectors():
-    layout = rot_surf_code(distance=3)
+    layout = rot_surface_code(distance=3)
     qubit_ids = {q: i for i, q in enumerate(layout.get_qubits())}
     model = NoiselessModel(qubit_ids)
     detectors = Detectors(layout.get_qubits(role="anc"), frame="1")
@@ -178,7 +178,7 @@ def test_memory_experiment_gauge_detectors():
 
 
 def test_repeated_s_experiment_gauge_detectors():
-    layout = rot_surf_code_rectangle(distance_z=4, distance_x=3)
+    layout = rot_surface_code_rectangle(distance_z=4, distance_x=3)
     set_trans_s(layout, "D1")
     qubit_ids = {q: i for i, q in enumerate(layout.get_qubits())}
     model = NoiselessModel(qubit_ids)
