+++ tests/layouts/test_util.py
@@ -1,10 +1,10 @@
 import pytest
 
-from surface_sim.layouts import rot_surf_code, set_coords
+from surface_sim.layouts import rot_surface_code, set_coords
 
 
 def test_set_coords():
-    layout = rot_surf_code(distance=3)
+    layout = rot_surface_code(distance=3)
 
     with pytest.raises(Exception):
         set_coords(layout)
