+++ tests/util/test_data_gen.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from surface_sim.layouts import rot_surf_code
+from surface_sim.layouts import rot_surface_code
 from surface_sim.util import sample_memory_experiment
 from surface_sim.models import NoiselessModel
 from surface_sim.experiments.rot_surface_code_xzzx import memory_experiment
@@ -8,7 +8,7 @@ from surface_sim import Detectors
 
 
 def test_sample_memory_experiment():
-    layout = rot_surf_code(distance=3)
+    layout = rot_surface_code(distance=3)
     qubit_ids = {q: i for i, q in enumerate(layout.get_qubits())}
     model = NoiselessModel(qubit_ids)
     detectors = Detectors(layout.get_qubits(role="anc"), frame="1")
